Rename paidbribe.py to paidbribeSprider.py
Scrapy crawler for scrapping recent 1000 posts from ww.ipaidbrbe.com. The site lays emphasis on the various kinds of Retail Corruption. This is kind of corruption that confronts ordinary citizens in their daily lives when they’re not able to avail of services they are legitimately entitled to from the government – getting a driver’s license, a birth certificate, registering a purchase of property.
The spider tries to scrape the following information : 
title
amount
name of department
transaction detail
number of views
city
date